trusted-firmware-m: trusted-firmware-m: bump to upstream TF-Mv1.2-RC2
Synchronize the trusted-firmware-m directory with the
corresponding upstream repository up to commit
368a25ff5e4925eed532fe7f6d03c3be9d8c4062
(TF-Mv1.2-RC2)

+++ trusted-firmware-m/platform/ext/common/tfm_hal_its.c
@@ -9,7 +9,7 @@
 #include "flash_layout.h"
 #include "tfm_hal_its.h"
 
-__WEAK void tfm_hal_its_fs_info(uint32_t *flash_area_addr, uint32_t *flash_area_size)
+__WEAK void tfm_hal_its_fs_info(uint32_t *flash_area_addr, size_t *flash_area_size)
 {
     if (!flash_area_addr || !flash_area_size) {
         return;
